add RAFT server state to wa_raft_info
Summary:
Add the state of the RAFT server to `wa_raft_info` for easy asynchronous
access even in cases where the RAFT server is busy serving a prior request.

The `gen_server` should still call the module's `terminate` callback even
when the server stops due to a crash so it is enough to just clear the
server state entry in the `terminate` callback.

Reviewed By: jaher

Differential Revision: D36020350

fbshipit-source-id: 85ebb9a698e5b9f04969228bcd8b72c735f4dfbd