add byte size limit to wa_raft_log:fold
Summary:
## Summary

When reading RAFT log entries, currently, we allow for a byte size limit to be used to prevent us from sending large chunks of data as a part of replication traffic. The way this byte size limitation is currently implemented, we read data on the leader and then truncate it before sending it over network, to followers.

This is suboptimal, because we end up utilizing large amounts of memory for no reason. To alleviate some memory usage within RAFT, we can use this limit when reading data and only read the log entries that we are going to replicate.

## `wa_raft_log_ets` Changes

In the ets-based implementation of RAFT, we use this limit in a standard fold and stop reading when we've accumulated log entries with a size >= to our byte limit.

## Notes

It is important to note that we've passed on the responsibility of limiting the size of replicated log entries to the provider, rather than RAFT itself. Implementations of `wa_raft` should keep this in mind when implementing new log providers.

--------

Reviewed By: hsun324

Differential Revision: D37245297

fbshipit-source-id: a74c6e0791563185d1dddc24d35f5391a380141e