witness in shard map and membership configuration
Summary:
Description:

To truly support witness replicas, the leader of the RAFT cluster must also know that a particular member of the cluster is expected to be a witness rather than a full member. The RAFT cluster knows about the members of the cluster by consulting the “membership” data that is stored as part of the RAFT cluster “config”. This config is replicated as part of the normal operation of the RAFT cluster and is saved in the RAFT storage.

Reviewed By: hsun324

Differential Revision: D37491063

fbshipit-source-id: fbaf9bd46a27d1a223ec45fec06f2f0ab8c11985