snapshot transfer to witness
Summary: Since RAFT witness replicas do not keep a database, we do not need to send the entire database to the witness replica when it is behind.  This will adjust the RAFT leader implementation to avoid sending a snapshot transfer that contains the entire database to a witness replica during follower catchup.

Reviewed By: hsun324

Differential Revision: D37530060

fbshipit-source-id: b698459386ddda78377e72681c945695009411ee