add RPC format with name
Summary:
Add a new RPC format that includes the service name of the sending
RAFT server. This is the first step towards making RAFT fully aware of
service names.

Adds an application flag that instructs RAFT servers to emit RPCs with
service names and adds a clause to all RAFT servers that downgrades
RPCs with service names to the equivalent non-service name RPC for
RPCs that do not yet support service names.

Reviewed By: jaher

Differential Revision: D38052066

fbshipit-source-id: 252fc24ac82e1f236ead626a5195c54c556e17a4