Add support for sending files in chunks
Summary:
Add the ability for transports to optionally return `continue` as the
status to indicate that the file was not completely sent within the call
to the transport provider. The transport worker will keep the file in
the queue and call the transport provider at a later time to continue
transport.

Reviewed By: jaher

Differential Revision: D38290597

fbshipit-source-id: 7c21c47fffb35040cf2f29bf7029c8648dbf99c9