Add queue server to RAFT partition
Summary:
The RAFT acceptor server is somewhat overloaded, performing operations
related to both accepting requests and dealing with pending requests. In
this way, the acceptor relies on server, which relies on storage, when
in turn replies on the acceptor, creating a dependency cycle which makes
supervision and startup difficult.

Factor out the functionality in the acceptor server that deals with
pending requests and queuing and place it in its own "queue" server.
Additionally, move macros related to the queue out of the RAFT application
header into the queue module itself to improve separation of concerns.

The queue server does not depend on any other service and can be started
first. This breaks the circular dependency between acceptor and storage.

This change also removes some legacy functionality that we did not
intent to support, namely, stale read with expected log index (which can
be implemented better by users themselves by accessing their storage
FSM implementation directly).

Reviewed By: jaher

Differential Revision: D38472278

fbshipit-source-id: 0902fbcdb57c4ea3eae96bacb141f1c6417f84a1