Avoid updating transport timestamp when info is not changed
Summary:
Do not set the last updated timestamp for a transport or transport file
if the info produced by the callback when updating a transport or
transport file info is unchanged. This prevents the regular transport
server scan for idle or stuck transports from resetting the idle timeout
each scan, resulting in unnaturally long delays before transports are
considered timed out.

Reviewed By: fnabulsi, jaher

Differential Revision: D39435399

fbshipit-source-id: 7c5c841d536d5fd0aea56257e7a0e8e700bd0594