Split out snapshot catchup from logs catchup
Summary:
Logs catchup and snapshot catchup do not necessarily need
to be handled in the same way. We still need some centralized
way to control the number of snapshots being transported at
a time and we want to avoid taking the snapshot from the
RAFT server process directly.

As a first step, split up logs atchup and snapshot catchup and
make snapshot catchup completely asynchronous. This will
make it simpler to make further adjustments to snapshot
transport and logs catchup without affecting the other.

Reviewed By: jaher

Differential Revision: D40321098

fbshipit-source-id: 3c967161849416c9be6efd5851bd22d44281f1b3