Remove override for log entry at zero
Summary:
The RAFT log is ideally always contiguous. Non-contiguous entries in the
RAFT log will be detected as corruption and could lead to crashes when
a leader attempts to send a heartbeat.

Originally, the override at log entry zero was intended to allow log
implementations to not need to consider special behavior for the log
entry at index zero; however, with the new RAFT log provider behaviour,
the log entry at index zero does not actually have any special behaviour
anymore but rather causes the RAFT log to be non-contiguous, leading
to potential crashes when the RAFT cluster is bootstrapping.

This change removes the special case for log entries at index zero and
adds some additional checks to the RAFT log server to protect the log
entry at index zero.

Note that after this change, in this implementation, the only time that a
log entry at index zero can be created is as a result of the `reset`
instruction with a position of `#raft_log_pos{index = 0, term = 0}`.
All other commands can only possibly manipulate log entries past
index zero.

Reviewed By: jaher

Differential Revision: D40777022

fbshipit-source-id: a03b95304ef487ae6adacbe46e8668d5b5b9796f