Split responsibility for handling RPCs and executing RPC requests
Summary:
Historically, changes to RPC formats in WARaft required that code to
support both the old and new formats for any particular RPC be
running for long enough for the full migration to occur. This is a
consequence of the implementation of the RAFT server matching
against exact RPCs and, therefore, expecting other RAFT servers to
send exactly the supported RPC formats.

This diff splits the responsibility for handling incoming RPCs and
unmarshalling them into a format supported by the local node and
the actual execution of the requested procedure call by splitting off
the concept of a procedure call from the RPC format.

After this diff, RAFT servers will execute procedures as the result
of receiving a procedure call which will always be expected to be
in a locally defined format while RPCs received from other nodes
will first be unmarshalled into a procedure. This splits the
responsibility for handling different RPC formats and supporting
forwards and backwards compatibility for RPC formats from the
implementations of the procedures themselves.

Reviewed By: jaher

Differential Revision: D40440531

fbshipit-source-id: 576479a8531269705f3709820d7df9930ec99d3c