Use record for options
Summary:
Since the options mapping is already intended to have a fixed size since
it must define all expected fields, we can just use a record instead.
Accesses against records are more efficient than accesses against a map.

This change is in support for multiple applications. As part of multi-app
support, shared services like transport will need to be able to quickly get
information about the RAFT servers that comprise a RAFT partition. The
plan is to place these option records into persistent term for quick access
as literals. *RAFT partition options are not expected to change during the
runtime of a node so GC impact is expected to be minimal.*

Reviewed By: jaher

Differential Revision: D42462132

fbshipit-source-id: 3a383d93f0f079dad9b31e02ad11021f91e38d98