Adjust storage provider interface for open and close
Summary:
Clean up the `wa_raft_storage` behaviour callbacks for opening
and closing databases (and snapshots), reducing the usage of
the internal `#raft_storage{}` record (which is the entire state of
the RAFT storage server).

Add a new `wa_raft_storage:storage_position` callback that is
expected to return the position of the most recent log entry
applied to the storage database. Generally, this function is used
at low frequency and so does not necessarily need to avoid I/O.

Reviewed By: jaher

Differential Revision:
D43521126

Privacy Context Container: L1141030

fbshipit-source-id: 1dc2992ad49f34f5c70fff41636abf89e966204e