Simplify storage snapshot interface
Summary:
Remove unnecessary `storage_delete_snapshot` callback because
all snapshots are defined to be directory trees and so deletion is
already generic. Refactor remaining storage callbacks so that
`#raft_storage{}` record is not leaked. Simplify storage snapshot
callbacks by passing in the full path to the snapshot directory instead
of the just the snapshot name.

Reviewed By: jaher

Differential Revision:
D43640109

Privacy Context Container: L1141030

fbshipit-source-id: 099719ec1f11f26393c521955e4f484b0a871402