Fix wa_raft_server_SUITE:persist_state
Summary:
OTP 26 changes the serialization of `term_to_binary`.

This diff fixes the load/store function to use the OTP 25 version of term_to_binary while the raft team finds a better longer term fix.

Note, the implementation of maps also changes in OTP 26. This results in indeterministic  ordering of key/value pairs retrieved from a map, which results in indeterministic ordering for term_to_binary.

We are storing the following:

```
StateList = [
        {current_term, CurrentTerm},
        {voted_for, VotedFor},
        {disable_reason, DisableReason}
],
```

`current_term` is a `non_neg_integer()`
`voted_for` is `undefined | node()`
`disable_reason` is a `term()` but it is `string() | undefined` in practice.

Therefore, we can make this change without breaking anything.

Differential Revision: D47976672

fbshipit-source-id: bf2c740a0aeb3ca74052f06825ede5b23be97f79