Fix pending log entries remaining after resignation
Summary:
Fix pending log entries remaining in the log view of the RAFT server
after the leader resigns due to eligibility or manual request to resign.
This fixes an issue where if the server becomes leader of a future term,
it may append log entries from an older term remaining in the pending
queue to the end of the log of a newer term, violating the log ordering
property.

Differential Revision: D48235959

fbshipit-source-id: e726a148ef54757bf243bbff7aff01dc6788fcd6