Simplify handling of log entries to append
Summary:
Simplify the handling of log entries to be appended by simplifying the
interface of `append_entries` and extracting code for handling heartbeats
duplicated between the `follower` and `witness` implementations.

Improve diagnostic logs generated by the RAFT server when a log mismatch
causes a heartbeat to be rejected by including the information that was
used to make the decision in the diagnostics.

Properly separate metrics between followers and witnesses generated
during heartbeat handling.

Differential Revision: D48524182

fbshipit-source-id: 975718bffd6dcfe76c00d54c21e85d21603e2266