Adjust ETS storage example to use unnamed tables
Summary:
Using a named table in the example ETS-based storage provider allows for
convenient access to the backing ETS table but causes issues when
creating and opening snapshots. Adjust the implementation to properly
track the storage position and use unnamed tables to avoid issues when
opening snapshots due to ETS's behavior of restoring the table's name
using using `file2tab`.

Differential Revision: D62262547

fbshipit-source-id: 51c4cc0cfed37502c486abe5d3ad85dffb736d07