Add an interface for log labeling
Summary:
Add an interface for labeling log entries with additional metadata as they are produced for the first time and added to the RAFT log.

The interface for labeling log entries is just:

wa_raft_label:new_label(LastLogLabel :: label(), NewCommand :: wa_raft_acceptor:command()) -> NewLogLabel :: label().

This API takes as input the label from the preceeding log entry and the payload for the log record being labelled (and just returns the label that needs to be attached).

This is a primitive for carrying over state across log entries (for example for labeling all log entries with a monotonically increasing timestamp or a hybrid logical clock).

Storage is required to persist the LogLabel (on each apply).

Differential Revision:
D63163719

Privacy Context Container: L1141030

fbshipit-source-id: 096798fce627958b16fdbfd28582ccf1c02af34e