Use an ETS-based queue for assigning files to transport senders
Summary:
Instead of instructing senders to send specific files, use an
ETS table as a queue listing the remaining files in an
ongoing transport. Transport senders will be notified of the
new transport and remove files from the queue. This avoids
a failure scenario where a stuck sender causes files to be
dropped from a transport.

Reviewed By: jaher

Differential Revision: D65612901

fbshipit-source-id: 896c210d5fb585fbaf211b85c535b49066537bbe