Add optional storage callback for creating empty snapshots for bootstrapping
Summary:
Add an optional callback to the `wa_raft_storage` behavior that
can be implemented to provide a way to construct a snapshot
of an empty storage state. This empty storage state can be used
to more reliably bootstrap RAFT partitions by simultaneously
constructing and loading the same empty storage state over all
replicas of a cluster when compared to the traditional method
of bootstrapping via promotion of a single leader.

Reviewed By: jaher

Differential Revision: D68859661

fbshipit-source-id: 76d8e863e4193034dee6f4a39096d811eb7f3f49