Only resign when ineligible
Summary:
Adjust the behavior of the RAFT server to no longer automatically
resign leadership when the `?RAFT_ELECTION_WEIGHT` is set
to zero.

Previously, the behavior of `?RAFT_ELECTION_WEIGHT` and
`?RAFT_LEADER_ELIGIBLE` overlapped here. Now only
`?RAFT_LEADER_ELIGIBLE` can cause the leader to resign and
prevent servers from becoming leader when explicitly requested.

Additionally, add a simpler and safer variant of the promotion
command that just causes a server to attempt to be elected. This
variant is intended to be used with the new empty snapshot-based
cluster bootstrap scheme.

Reviewed By: jaher

Differential Revision: D70811021

fbshipit-source-id: 0fc0502333866f9dccf4c26d6944b10a73e9faa4