Simplify election voting
Summary:
In the standard implementation of the RAFT protocol, `false` votes are
only necessary to notify candidates of newer terms. In the current
implementation, `RequestVote` RPCs with older terms are already
intercepted and a `NotifyTerm` RPC sent. As such, `false` votes
serve no purpose. This change avoids sending those votes.

In addition, refactor the voting logic to remove duplication between
the follower and witness state implementations.

Reviewed By: jaher

Differential Revision: D72195783

fbshipit-source-id: 570356314d3f15d66c27664d8ab558549fea8850