have wa_raft_storage:make_empty_snapshot return last applied log position
Summary:
Instead of forking the logic for witnesses and full replicas we'll instead create an empty snapshot with the config and the last applied log position and transfer it normally similar to the full replicas.
This diff just makes it so `wa_raft_storage:make_empty_snapshot` returns the last applied log position similarly to `wa_raft_storage:create_snapshot`.

Reviewed By: jaher

Differential Revision: D72344559

fbshipit-source-id: 6e8617b239d56c6cab04eced37457093a752de94