Disallow configuration changes during promotion
Summary:
Remove the ability for the promotion command to change the current config.

This change means that the promotion command alone is no longer sufficient to
bootstrap a RAFT cluster. Bootstrapping a new cluster now requires the bootstrap
command to be run on at least a quorum of the members in the provided config or
a bootstrap command followed by a promotion to leader.

Reviewed By: jaher

Differential Revision:
D72425782

Privacy Context Container: L1141030

fbshipit-source-id: 63b0490384aa278100063e3fa6d5ac1a6950a411