Use follower last applied index for handover restriction
Summary:
Add a new requirement for a peer to be considered a good candidate
to receive a handover. Peers whose storage states are missing a
significant number of applies when compared to the log state of the
leader will no longer be considered a good handover candidate.

This feature requires information about follower storage states to
be provided to the leader in `AppendEntriesResponse` RPCs and so
includes an API change. No special deployment steps are required
as the change is forwards and backwards compatible.

Reviewed By: jaher

Differential Revision:
D72591497

Privacy Context Container: L1141030

fbshipit-source-id: f31da13c0737ff5bb18c21051303fd763bfd26be