Properly track commit index
Summary:
Rather than setting `CommitIndex` equal to the last applied log
entry on leaders and ignoring `CommitIndex` on followers, actually
store the known commit index in the `commit_index` field of the
RAFT state.

This also changes restrictions on successive membership changes
and heartbeat truncation to be correctly based off of the known
commit index instead of the last applied index.

Reviewed By: jaher

Differential Revision: D73948079

fbshipit-source-id: 1d292a9ca132be8a2d76f96b9ef8c6b30f1e9391