Add no_dialyzer_attribute linter
Summary:
[Dialyzer](https://www.erlang.org/doc/apps/dialyzer/dialyzer_chapter.html#introduction) is a static analysis tool that identifies software discrepancies, such as
definite type errors, code that is unreachable because of programming error, and
unnecessary tests in single Erlang modules or an entire codebase.

The tool provides [a mechanism](https://www.erlang.org/doc/apps/dialyzer/dialyzer.html#module-requesting-or-suppressing-warnings-in-source-files) for turning off warnings in a module via the `-dialyzer` attribute. This approach is not recommended, as it can lead to errors being missed. Instead, it is recommended to fix the errors instead of suppressing them.

Reviewed By: alanz

Differential Revision: D74791045

fbshipit-source-id: d7113f456222dfb6e5be281826defef2976c3293