Move commit batching logic from log to server
Summary:
Commit batching is more a detail of the operation of the RAFT server than of the log. Moving this
to the server implementation will allow for more fine grained control of the actual log entries
contained within the batch.

This fixes a discrepancy where labels were being assigned to log entries within the commit
batch but the label was not rolled back if the batch failed to append.

Reviewed By: jaher

Differential Revision: D75103930

fbshipit-source-id: f563490235097d606f23d38a056ea62a57ea3740