Move partial append truncation logic to server
Summary:
Simplify log append logic by moving handling of truncation (which was previously
duplicated) to the RAFT server. This also fixes a potential issue where a heartbeat
with partially matching log terms can cause the RAFT log implementation to
attempt to truncate log entries past the RAFT server local commit index which the
RAFT log is not aware of.

Reviewed By: jaher

Differential Revision: D75886646

fbshipit-source-id: 45edeb6e8be24d604640aed3ecf551385c51b252