Only reply to incomplete commit requests when next leader is known
Summary:
Defer replying the incomplete commit requests or commit requests made during handovers until
after the leader of a new term is known or the local
replica regains leader by being elected again or
because a pending handover times out or fails. This reduces the number of cases where requests are
incorrectly rejected or replies are made early.

Reviewed By: jaher

Differential Revision: D76849186

fbshipit-source-id: b0ca4b34e382321d1d7245ace9e9a7169909f912