Avoid file server for more IO operations
Summary: By default, IO operations are serialised through the file server, which can add significant overhead. We can bypass this using the `raw` option or the `prim_file` module, depending on context. `wa_raft` already uses this in many places, but here we enable it in a few more to further boost performance.

Reviewed By: hsun324

Differential Revision: D78900618

fbshipit-source-id: 597baa97a66bb803f4f41f307d6e9ca2a011c507