Add live status for partitions
Summary:
Add a new "live" status for partitions that can be queried from `wa_raft_info` that represents whether a partition
believes it is an active and good participant in its current RAFT cluster. This status differs from "stale" as it is
not intended to indicate anything about whether the underlying FSM is in a good state.

Reviewed By: SarahDesouky

Differential Revision: D80120378

fbshipit-source-id: aee12c87d161a7d874eee72306ed1dde088b1f87