Update server queue length without requiring staleness change
Summary:
Introduce a separate info method for updating the server queue length and call
it unconditionally whenever we are reevaluating the liveness and staleness of the
RAFT server to improve the accuracy of the value.

This fixes an issue where the queue length value can get stuck if the RAFT server
stops receiving RPCs (for example if a custom distribution implementation withholds
RPCs when the queue is too large). Also fixes an issue where the queue length was
always being stored with the default name of the RAFT partition.

Reviewed By: jaher, SarahDesouky

Differential Revision:
D86316661

Privacy Context Container: L1384697

fbshipit-source-id: c1c91dea7dd5a1914565803bbfb01a168a104469