Add support for non-member replication participants
Summary:
Add support for non-member replicas to be added to the RAFT config to participate
in replication. This functionality can be used to add additional replicas to a cluster that do not
participate in quorum decisions but do receive replication traffic similar to normal replicas.

This change also cleans up and merges the previous two commands that both affected the
cluster configuration, merging `adjust_membership` and `refresh_config` into a single
`adjust_config` command.

Reviewed By: SarahDesouky

Differential Revision: D87011904

fbshipit-source-id: 750f3bc2c1d3ab25f9b9de06d229535570c2a26d