Avoid triggering poor Dialyzer performance
Summary:
There are dark corners of Dialyzer (some are reworked and improved in OTP 28), but certain rare combinations of types can trigger extremely bad performance. That was the case here, but it wasn't initially detected due to incrementality masking the impact somewhat. We rework the opaque usage here to avoid triggering that until all consumption has moved to a Dialyzer version which handles this more gracefully.

I tried to find a way to achieve this whilst keeping both the opaque and the two very similar maps, but failed, so one of those had to give. That's disappointing, but the guarantees given by the tooling for `-opaque` are quite weak anyway. Recent typing work should mean a better solution is possible in future.

Reviewed By: jaher

Differential Revision:
D87928250

Privacy Context Container: L1384697

fbshipit-source-id: ee038b99211ef594af9027bdf0a7cff039e70e27