Enable unused_macro linter in CI (prod + tests)
Summary:
The linter is enabled by default in ELP, and warnings appear in the IDE.
But the same linter is not explicitly added to the CI configuration, so warnings are not reported on diffs or in Contlint.

This diff enables the linter, aligning the IDE and the CI experience. It also makes it possible to track violations via ContLint.

Violations are solved via:

```
elp lint --project /data/sandcastle/boxes/fbsource/whatsapp/server --diagnostic-filter unused_macro --no-stream --include-tests --format json --apply-fix --in-place --one-shot
```

Followed by manual tweaking.

Differential Revision: D89378624

fbshipit-source-id: 3ca8876d4f0d2a316aa446049b2fe1dfeffbdd9d